이미지가 등록되어 있지 않은 경우 NPE 발생 문제 해결

+++ ward-shop/src/main/java/com/ward/wardshop/goods/service/ProductService.java
@@ -112,6 +112,8 @@ public class ProductService {
     }
 
     private void deleteImage(Product product) throws IOException {
-        imageManager.deleteImg(product.getImageResource());
+        if (Objects.nonNull(product.getImageResource())) {
+            imageManager.deleteImg(product.getImageResource());
+        }
     }
 }
